Added new simplified auth management file

+++ ThunderED/Json/AllianceData.cs
@@ -8,16 +8,16 @@ namespace ThunderED.Json
         {
             public string name { get; set; }
             public string ticker { get; set; }
-            public int creator_id { get; set; }
-            public int creator_corporation_id { get; set; }
-            public int executor_corporation_id { get; set; }
+            public long creator_id { get; set; }
+            public long creator_corporation_id { get; set; }
+            public long executor_corporation_id { get; set; }
             public DateTime date_founded { get; set; }
-            public int? faction_id { get; set; }
+            public long? faction_id { get; set; }
         }
 
         internal class AllianceIDLookup
         {
-            public int[] alliance;
+            public long[] alliance;
         }
 
         internal class FactionData
